@@ -3,11 +3,11 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'com.novoda.bintray-release'//添加
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -37,17 +37,17 @@ configurations.all {
 }
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
     implementation files('libs/library-listviewanimations.jar')
     implementation files('libs/nineoldandroids-2.4.0.jar')
 }
 //添加
 publish {
-    artifactId = 'viewtools'
+    artifactId = 'viewtoolsX'
     userOrg = 'zhengzaihong'
     groupId = 'com.zzh'
-    uploadName = "viewtool"
+    uploadName = "viewtoolX"
     publishVersion = '0.1.93'
     desc = 'this quick tools make your app has kinds of view tools'
     website = 'https://github.com/zhengzaihong/MyViews.git'
