@@ -3,8 +3,8 @@ package dz.solc.viewtool.view.toolview;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 
