@@ -2,7 +2,7 @@ package dz.solc.viewtool.view.viewpager;
 
 import android.content.Context;
 import android.graphics.PointF;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
